requirements.txt
ccxt==4.2.85
Flask==3.0.2
python-dotenv==1.0.0
requests==2.31.0